python/packages/corlinman-skills-registry/pyproject.toml
[project]
name = "corlinman-skills-registry"
version = "0.1.0"
description = "Skill registry: parses openclaw-style SKILL.md files (YAML frontmatter + Markdown body) and checks runtime requirements."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "corlinman contributors" }]

dependencies = [
    "pydantic>=2.10",
    "PyYAML>=6.0",
    "structlog>=25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corlinman_skills_registry"]
